LayoutTests/js/dom/dfg-ensure-array-storage-on-window-expected.txt

Tests that passing the global object to an array access that will arrayify to ArrayWithArrayStorage doesn't break things.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(w) is NaN
PASS successfullyParsed is true

TEST COMPLETE
